requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI and LLM
google-generativeai==0.3.2
langchain==0.1.4
langchain-google-genai==0.0.6

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Embeddings and Vector Operations
sentence-transformers==2.3.1
numpy==1.24.3

# Environment Management
python-dotenv==1.0.0

# Templating
jinja2==3.1.3
